+++ apps/infrastructure/src/modules/chatbot/lambda_chatbot.tf
@@ -1,11 +1,6 @@
 locals {
   lambda_env_variables = {
     AUTH_COGNITO_USERPOOL_ID           = var.cognito_user_pool.id
-    CHB_AWS_BEDROCK_EMBED_REGION       = "eu-central-1"
-    CHB_AWS_BEDROCK_LLM_REGION         = var.aws_chatbot_region
-    CHB_AWS_BEDROCK_RERANKER_REGION    = "eu-central-1"
-    CHB_AWS_GUARDRAIL_ID               = awscc_bedrock_guardrail.guardrail.guardrail_id
-    CHB_AWS_GUARDRAIL_VERSION          = awscc_bedrock_guardrail_version.guardrail.version
     CHB_AWS_S3_BUCKET                  = module.s3_bucket_llamaindex.s3_bucket_id
     CHB_AWS_SSM_GOOGLE_API_KEY         = module.google_api_key_ssm_parameter.ssm_parameter_name
     CHB_AWS_SSM_GOOGLE_SERVICE_ACCOUNT = module.google_service_account_ssm_parameter.ssm_parameter_name
@@ -27,28 +22,30 @@ locals {
     # Be extremely careful when changing the provider
     # both the generation and the embedding models would be changed
     # embeddings size change would break the application and requires reindexing
-    CHB_PROVIDER               = var.models.provider
-    CHB_QUERY_TABLE_PREFIX     = local.prefix
-    CHB_REDIS_URL              = "redis://${module.nlb.dns_name}:${var.ecs_redis.port}"
-    CHB_RERANKER_ID            = var.models.reranker
-    CHB_USE_PRESIDIO           = "True"
-    CHB_WEBSITE_URL            = "https://${var.dns_domain_name}"
-    CORS_DOMAINS               = var.environment == "dev" ? jsonencode(["https://www.${var.dns_domain_name}", "https://${var.dns_domain_name}", "http://localhost:3000"]) : jsonencode(["https://www.${var.dns_domain_name}", "https://${var.dns_domain_name}"])
-    ENVIRONMENT                = var.environment
-    LLAMA_INDEX_CACHE_DIR      = "/tmp"
-    LOG_LEVEL                  = "INFO"
-    NLTK_DATA                  = "_static/nltk_cache/"
-    TIKTOKEN_CACHE_DIR         = "/tmp/tiktoken"
-    CHB_AWS_SSM_STRAPI_API_KEY = "/chatbot/chb_strapi_api_key"
+    CHB_PROVIDER                          = var.models.provider
+    CHB_QUERY_TABLE_PREFIX                = local.prefix
+    CHB_REDIS_URL                         = "redis://${module.nlb.dns_name}:${var.ecs_redis.port}"
+    CHB_RERANKER_ID                       = var.models.reranker
+    CHB_USE_PRESIDIO                      = "True"
+    CHB_WEBSITE_URL                       = "https://${var.dns_domain_name}"
+    CHB_AWS_S3_BUCKET_NAME_STATIC_CONTENT = var.s3_bucket_name_static_content
+    CORS_DOMAINS                          = var.environment == "dev" ? jsonencode(["https://www.${var.dns_domain_name}", "https://${var.dns_domain_name}", "http://localhost:3000"]) : jsonencode(["https://www.${var.dns_domain_name}", "https://${var.dns_domain_name}"])
+    ENVIRONMENT                           = var.environment
+    LLAMA_INDEX_CACHE_DIR                 = "/tmp"
+    LOG_LEVEL                             = "INFO"
+    NLTK_DATA                             = "_static/nltk_cache/"
+    TIKTOKEN_CACHE_DIR                    = "/tmp/tiktoken"
+    CHB_AWS_SSM_STRAPI_API_KEY            = "/chatbot/chb_strapi_api_key"
+    CHB_AWS_SQS_QUEUE_EVALUATE_NAME       = aws_sqs_queue.chatbot_evaluate_queue.name
   }
 }
 
-# Diretta implementazione delle risorse che il modulo lambda creerebbe
+
 resource "aws_lambda_function" "chatbot_lambda" {
   function_name = "${local.prefix}-api-lambda"
   description   = "Lambda function running APIs of the Developer Portal Chatbot"
 
-  image_uri    = "${module.ecr.repository_url}:latest"
+  image_uri    = "${module.ecr["chatbot"].repository_url}:latest"
   package_type = "Image"
 
   timeout       = local.lambda_timeout
@@ -211,8 +208,47 @@ resource "aws_lambda_permission" "allow_eventbridge" {
 }
 
 # IAM Policy Resources
-resource "aws_iam_policy" "lambda_s3_bedrock_policy" {
-  name = "chatbot-${var.environment}-api-lambda-s3-bedrock"
+
+resource "aws_iam_policy" "lambda_chatbot_ecr_access" {
+
+  name        = "chatbot-lambda-ecr-access"
+  description = "Allow Lambda to pull images from ECR chatbot repositories"
+  policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Effect = "Allow"
+        Action = [
+          "ecr:ListTagsForResource",
+          "ecr:ListImages",
+          "ecr:GetRepositoryPolicy",
+          "ecr:GetLifecyclePolicyPreview",
+          "ecr:GetLifecyclePolicy",
+          "ecr:GetDownloadUrlForLayer",
+          "ecr:GetAuthorizationToken",
+          "ecr:DescribeRepositories",
+          "ecr:DescribeImages",
+          "ecr:DescribeImageScanFindings",
+          "ecr:BatchGetImage",
+          "ecr:BatchCheckLayerAvailability",
+        ]
+        Resource = "*"
+      },
+      {
+        Effect = "Allow"
+        Action = [
+          "ecr:GetAuthorizationToken"
+        ]
+        Resource = "*"
+      }
+    ]
+  })
+}
+
+
+
+resource "aws_iam_policy" "lambda_s3_chatbot_policy" {
+  name = "chatbot-${var.environment}-api-lambda-s3"
   policy = jsonencode({
     Version = "2012-10-17"
     Statement = [
@@ -227,20 +263,6 @@ resource "aws_iam_policy" "lambda_s3_bedrock_policy" {
         Effect   = "Allow"
         Action   = "s3:*Object"
         Resource = ["${module.s3_bucket_llamaindex.s3_bucket_arn}/*", "${module.s3_bucket_kb.s3_bucket_arn}/*"]
-      },
-      {
-        Sid    = "BedrockPermissions"
-        Effect = "Allow"
-        Action = [
-          "bedrock:ApplyGuardrail",
-          "bedrock:ListGuardrails",
-          "bedrock:GetGuardrail",
-          "bedrock:InvokeModel",
-          "bedrock:InvokeModelWithResponseStream",
-          "bedrock:ListFoundationModels",
-          "bedrock:Rerank"
-        ]
-        Resource = "*"
       }
     ]
   })
@@ -310,17 +332,17 @@ resource "aws_iam_policy" "chatbot_monitor_queue" {
     Statement = [
       {
         Effect   = "Allow"
-        Action   = "sqs:SendMessage"
-        Resource = aws_sqs_queue.chatbot_monitor_queue.arn
+        Action   = ["sqs:SendMessage", "sqs:GetQueueUrl"]
+        Resource = aws_sqs_queue.chatbot_evaluate_queue.arn
       }
     ]
   })
 }
 
 # IAM Role Policy Attachments
-resource "aws_iam_role_policy_attachment" "lambda_s3_bedrock_policy_attachment" {
+resource "aws_iam_role_policy_attachment" "lambda_s3_chatbot_policy_attachment" {
   role       = aws_iam_role.lambda_role.name
-  policy_arn = aws_iam_policy.lambda_s3_bedrock_policy.arn
+  policy_arn = aws_iam_policy.lambda_s3_chatbot_policy.arn
 }
 
 resource "aws_iam_role_policy_attachment" "lambda_dynamodb_policy_attachment" {
@@ -336,6 +358,12 @@ resource "aws_iam_role_policy_attachment" "lambda_ssm_policy_attachment" {
 resource "aws_iam_role_policy_attachment" "lambda_logs_policy_attachment" {
   role       = aws_iam_role.lambda_role.name
   policy_arn = "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
+}
+
+
+resource "aws_iam_role_policy_attachment" "lambda_chatbot_ecr_access_attach" {
+  role       = aws_iam_role.lambda_role.name
+  policy_arn = aws_iam_policy.lambda_chatbot_ecr_access.arn
 }
 
 resource "aws_iam_role_policy_attachment" "lambda_vpc_policy_attachment" {