+++ apps/nextjs-website/src/app/layout.tsx
@@ -14,7 +14,6 @@ import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import '@/styles/globals.css';
-import '@/polyfill';
 import ThemeRegistry from './ThemeRegistry';
 import { getProducts } from '@/lib/api';
 import SiteFooter from '@/components/atoms/SiteFooter/SiteFooter';
@@ -28,7 +27,6 @@ import { Titillium_Web } from 'next/font/google';
 import NextIntlContext from '@/components/atoms/NextIntlContext/NextIntlContext';
 import ChatbotProvider from '@/components/organisms/ChatbotProvider/ChatbotProvider';
 import { ErrorBoundary } from 'next/dist/client/components/error-boundary';
-import ErrorPage from '@/components/atoms/ErrorPage/ErrorPage';
 import Error from './error';
 
 // TODO: remove PREVIOUS_MATOMO_TAG_MANAGER_SCRIPT script, usePreviousScript when the migration to the new tag manager is completed
@@ -89,10 +87,6 @@ const titilliumWeb = Titillium_Web({
 export const metadata: Metadata = {
   metadataBase: new URL(baseUrl),
 };
-
-export function generateStaticParams() {
-  return [{ locale: 'en' }, { locale: 'it' }];
-}
 
 export default async function RootLayout({
   // Layouts must accept a children prop.