+++ apps/nextjs-website/src/components/molecules/SubscribeToWebinar/SubscribeToWebinar.tsx
@@ -1,5 +1,5 @@
 'use client';
-import React, { useCallback, useEffect } from 'react';
+import React, { useCallback, useEffect, Suspense } from 'react';
 import { useState } from 'react';
 import SubscribeButton from '../../atoms/SubscribeButton/SubscribeButton';
 import { useTranslations } from 'next-intl';
@@ -7,6 +7,7 @@ import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 import { WebinarState } from '@/helpers/webinar.helpers';
 import { subscribeToWebinar, unsubscribeToWebinar } from '@/lib/webinarApi';
 import { useUser } from '@/helpers/user.helper';
+import Spinner from '../../atoms/Spinner/Spinner';
 
 export type SubscribeButtonProps = {
   webinarSlug?: string;
@@ -14,14 +15,16 @@ export type SubscribeButtonProps = {
   setIsSubscribed: (isSubscribed: boolean) => null;
   handleErrorMessage?: (message: string) => null;
   webinarState: WebinarState;
+  textColor?: string;
 };
 
-const SubscribeToWebinar = ({
+const SubscribeToWebinarContent = ({
   webinarSlug,
   isSubscribed,
   setIsSubscribed,
   handleErrorMessage,
   webinarState,
+  textColor = 'white',
 }: SubscribeButtonProps) => {
   const t = useTranslations('webinar');
   const [isLoading, setIsLoading] = useState(false);
@@ -148,7 +151,16 @@ const SubscribeToWebinar = ({
       onSubscribe={onSubscribeClick}
       onCancelSubscription={onUnsubscribe}
       subscribeLabel={subscribeLabelMap[webinarState]}
+      textColor={textColor}
     />
+  );
+};
+
+const SubscribeToWebinar = (props: SubscribeButtonProps) => {
+  return (
+    <Suspense fallback={<Spinner />}>
+      <SubscribeToWebinarContent {...props} />
+    </Suspense>
   );
 };