+++ apps/nextjs-website/src/lib/cmsApi.ts
@@ -22,9 +22,7 @@ import { makeApiDataListProps } from './strapi/makeProps/makeApiDataList';
 import { fetchApiDataList } from './strapi/fetches/fetchApiDataList';
 import { fetchProducts } from '@/lib/strapi/fetches/fetchProducts';
 import { makeProductsProps } from './strapi/makeProps/makeProducts';
-import { fetchGuideListPages } from './strapi/fetches/fetchGuideListPages';
 import { makeGuideListPagesProps } from './strapi/makeProps/makeGuideListPages';
-import { fetchGuide, fetchGuides } from './strapi/fetches/fetchGuides';
 import { makeGuidesProps } from './strapi/makeProps/makeGuides';
 import { fetchOverviews } from '@/lib/strapi/fetches/fetchOverviews';
 import { makeOverviewsProps } from '@/lib/strapi/makeProps/makeOverviews';
@@ -32,7 +30,6 @@ import { fetchTutorialListPages } from './strapi/fetches/fetchTutorialListPages'
 import { makeTutorialListPagesProps } from './strapi/makeProps/makeTutorialListPages';
 import { fetchUrlReplaceMap } from './strapi/fetches/fetchUrlReplaceMap';
 import { makeUrlReplaceMap } from './strapi/makeProps/makeUrlReplaceMap';
-import { withCache, getCacheKey } from './cache';
 import { makeReleaseNotesProps } from '@/lib/strapi/makeProps/makeReleaseNotes';
 import { fetchReleaseNote } from '@/lib/strapi/fetches/fetchReleaseNotes';
 import {
@@ -40,11 +37,15 @@ import {
   makeSolution as makeSolutionS3,
   makeReleaseNote as makeReleaseNoteS3,
 } from '@/helpers/makeS3Docs.helpers';
-// import { makeGuide, makeReleaseNote } from '@/helpers/makeDocs.helpers';
 import { secrets } from '@/config';
 import { fetchWebinarCategories } from '@/lib/strapi/fetches/fetchWebinarCategories';
 import { makeWebinarCategoriesProps } from '@/lib/strapi/makeProps/makeWebinarCategories';
-import { JsonMetadata } from '@/helpers/s3Metadata.helpers';
+import {
+  fetchResponseFromCDN,
+  JsonMetadata,
+} from '@/helpers/s3Metadata.helpers';
+import { StrapiGuideListPages } from './strapi/codecs/GuideListPagesCodec';
+import { StrapiGuides } from './strapi/codecs/GuidesCodec';
 
 // a BuildEnv instance ready to be used
 const buildEnv = pipe(
@@ -56,199 +57,89 @@ const buildEnv = pipe(
   })
 );
 
-const CACHE_EXPIRY_IN_SECONDS = 900; // 15 minutes in seconds
-
 export const getHomepageProps = async () => {
-  return withCache(
-    getCacheKey('getHomepageProps'),
-    async () => {
-      const strapiHomepage = await fetchHomepage(buildEnv);
-      return makeHomepageProps(strapiHomepage);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiHomepage = await fetchHomepage(buildEnv);
+  return makeHomepageProps(strapiHomepage);
 };
 
 export const getWebinarsProps = async () => {
-  return withCache(
-    getCacheKey('getWebinarsProps'),
-    async () => {
-      const strapiWebinars = await fetchWebinars(buildEnv);
-      return makeWebinarsProps(strapiWebinars);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiWebinars = await fetchWebinars(buildEnv);
+  return makeWebinarsProps(strapiWebinars);
 };
 
 export const getProductsProps = async () => {
-  return withCache(
-    getCacheKey('getProductsProps'),
-    async () => {
-      const strapiProducts = await fetchProducts(buildEnv);
-      return makeProductsProps(strapiProducts);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiProducts = await fetchProducts(buildEnv);
+  return makeProductsProps(strapiProducts);
 };
 
 export const getWebinarCategoriesProps = async () => {
-  return withCache(
-    getCacheKey('getWebinarCategoriesProps'),
-    async () => {
-      const strapiWebinarCategories = await fetchWebinarCategories(buildEnv);
-      return makeWebinarCategoriesProps(strapiWebinarCategories);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiWebinarCategories = await fetchWebinarCategories(buildEnv);
+  return makeWebinarCategoriesProps(strapiWebinarCategories);
 };
 
 export const getTutorialsProps = async () => {
-  return withCache(
-    getCacheKey('getTutorialsProps'),
-    async () => {
-      const strapiTutorials = await fetchTutorials(buildEnv);
-      return makeTutorialsProps(strapiTutorials);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiTutorials = await fetchTutorials(buildEnv);
+  return makeTutorialsProps(strapiTutorials);
 };
 
 export const getTutorialListPagesProps = async () => {
-  return withCache(
-    getCacheKey('getTutorialListPagesProps'),
-    async () => {
-      const strapiTutorialListPages = await fetchTutorialListPages(buildEnv);
-      return makeTutorialListPagesProps(strapiTutorialListPages);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiTutorialListPages = await fetchTutorialListPages(buildEnv);
+  return makeTutorialListPagesProps(strapiTutorialListPages);
 };
 
 export const getQuickStartGuidesProps = async () => {
-  return withCache(
-    getCacheKey('getQuickStartGuidesProps'),
-    async () => {
-      const strapiQuickStartGuides = await fetchQuickStartGuides(buildEnv);
-      return makeQuickStartGuidesProps(strapiQuickStartGuides);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiQuickStartGuides = await fetchQuickStartGuides(buildEnv);
+  return makeQuickStartGuidesProps(strapiQuickStartGuides);
 };
 
 export const getUrlReplaceMapProps = async () => {
-  return withCache(
-    getCacheKey('getUrlReplaceMapProps'),
-    async () => {
-      const strapiUrlReplaceMap = await fetchUrlReplaceMap(buildEnv);
-      return makeUrlReplaceMap(strapiUrlReplaceMap);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiUrlReplaceMap = await fetchUrlReplaceMap(buildEnv);
+  const processed = makeUrlReplaceMap(strapiUrlReplaceMap);
+  return processed;
 };
 
 export const getApiDataListPagesProps = async () => {
-  return withCache(
-    getCacheKey('getApiDataListPagesProps'),
-    async () => {
-      const strapiApiDataListPages = await fetchApiDataListPages(buildEnv);
-      return makeApiDataListPagesProps(strapiApiDataListPages);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiApiDataListPages = await fetchApiDataListPages(buildEnv);
+  return makeApiDataListPagesProps(strapiApiDataListPages);
 };
 
 export const getApiDataProps = async () => {
-  return withCache(
-    getCacheKey('getApiDataProps'),
-    async () => {
-      const strapiApiDataList = await fetchApiDataList(buildEnv);
-      return await makeApiDataListProps(strapiApiDataList);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiApiDataList = await fetchApiDataList(buildEnv);
+  return await makeApiDataListProps(strapiApiDataList);
 };
 
 export const getCaseHistoriesProps = async () => {
-  return withCache(
-    getCacheKey('getCaseHistoriesProps'),
-    async () => {
-      const strapiCaseHistories = await fetchCaseHistories(buildEnv);
-      return makeCaseHistoriesProps(strapiCaseHistories);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiCaseHistories = await fetchCaseHistories(buildEnv);
+  return makeCaseHistoriesProps(strapiCaseHistories);
 };
 
 export const getSolutionsProps = async () => {
-  return withCache(
-    getCacheKey('getSolutionsProps'),
-    async () => {
-      const strapiSolutions = await fetchSolutions(buildEnv);
-      return makeSolutionsProps(strapiSolutions);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiSolutions = await fetchSolutions(buildEnv);
+  return makeSolutionsProps(strapiSolutions);
 };
 
 export const getSolutionListPageProps = async () => {
-  return withCache(
-    getCacheKey('getSolutionListPageProps'),
-    async () => {
-      const strapiSolutionListPage = await fetchSolutionListPage(buildEnv);
-      return makeSolutionListPageProps(strapiSolutionListPage);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiSolutionListPage = await fetchSolutionListPage(buildEnv);
+  return makeSolutionListPageProps(strapiSolutionListPage);
 };
 
 export const getOverviewsProps = async () => {
-  return withCache(
-    getCacheKey('getOverviewsProps'),
-    async () => {
-      const strapiOverviews = await fetchOverviews(buildEnv);
-      return makeOverviewsProps(strapiOverviews);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
+  const strapiOverviews = await fetchOverviews(buildEnv);
+  return makeOverviewsProps(strapiOverviews);
 };
 
 export const getGuideListPagesProps = async () => {
-  return withCache(
-    getCacheKey('getGuideListPagesProps'),
-    async () => {
-      const strapiGuideList = await fetchGuideListPages(buildEnv);
-      return makeGuideListPagesProps(strapiGuideList);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
-};
-
-export const getGuidesProps = async () => {
-  return withCache(
-    getCacheKey('getGuidesPropsCache'),
-    async () => {
-      const strapiGuides = await fetchGuides(buildEnv);
-      return makeGuidesProps(strapiGuides);
-    },
-    CACHE_EXPIRY_IN_SECONDS
-  );
-};
-
-export const getGuides = async () => {
-  const strapiGuides = await fetchGuides(buildEnv);
-  return makeGuidesProps(strapiGuides);
+  const strapiGuideList = (await fetchResponseFromCDN(
+    'synced-guide-list-pages-response.json'
+  )) as StrapiGuideListPages | undefined;
+  return strapiGuideList ? makeGuideListPagesProps(strapiGuideList) : [];
 };
 
 export const getGuideProps = async (
   guidePaths: ReadonlyArray<string>,
   productSlug: string
 ) => {
-  const strapiGuides = await fetchGuide(guidePaths[0], productSlug)(buildEnv);
-  if (!strapiGuides || strapiGuides.data.length < 1) {
-    // eslint-disable-next-line functional/no-throw-statements
-    throw new Error('Failed to fetch data');
-  }
-  const guide = makeGuidesProps(strapiGuides)[0];
+  const guide = await getGuidePageProps(guidePaths[0], productSlug);
   return await makeGuideS3({ guideDefinition: guide, guidePaths });
 };
 
@@ -256,13 +147,22 @@ export const getGuidePageProps = async (
   guideSlug: string,
   productSlug: string
 ) => {
-  const strapiGuides = await fetchGuide(guideSlug, productSlug)(buildEnv);
-  if (!strapiGuides || strapiGuides.data.length < 1) {
+  // TODO: restore this when Strapi will manage guides metadata
+  const strapiGuides = (await fetchResponseFromCDN(
+    'synced-guides-response.json'
+  )) as StrapiGuides | undefined;
+  // eslint-disable-next-line functional/no-expression-statements
+  const guides = strapiGuides ? makeGuidesProps(strapiGuides) : [];
+  const guide = guides.filter(
+    (g) => g.guide.slug === guideSlug && g.product.slug === productSlug
+  )[0];
+
+  if (!guide) {
     // eslint-disable-next-line functional/no-throw-statements
-    throw new Error('Failed to fetch data');
+    throw new Error('Failed to fetch guide data');
   }
-  const guidesProps = makeGuidesProps(strapiGuides);
-  return guidesProps[0];
+
+  return guide;
 };
 
 export const getSolutionProps = async (