+++ apps/strapi-cms/src/api/api-data/controllers/api-data.ts
@@ -1,7 +1,0 @@
-/**
- * api-data controller
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreController('api::api-data.api-data');