+++ apps/strapi-cms/src/api/case-history/controllers/case-history.ts
@@ -1,7 +1,0 @@
-/**
- * case-history controller
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreController('api::case-history.case-history');