+++ apps/strapi-cms/src/api/guide/controllers/guide.ts
@@ -1,7 +1,0 @@
-/**
- * guide controller
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreController('api::guide.guide');