+++ apps/strapi-cms/src/api/quickstart-guide-item/routes/quickstart-guide-item.ts
@@ -1,9 +1,0 @@
-/**
- * quickstart-guide-item router
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreRouter(
-  'api::quickstart-guide-item.quickstart-guide-item'
-);