+++ apps/strapi-cms/src/api/solution/services/solution.ts
@@ -1,7 +1,0 @@
-/**
- * solution service
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreService('api::solution.solution');